fix(youtube): Cleaned up and moved YouTube code, switch to API key
This switches our authentication to use a developerKey as a static auth
method instead of the OAuth flow. The OAuth flow code has been kept as a
legacy comment as it will have a future use in the website backend flow
for users. This should enable us to start running updates and static
generation as part of a CI process.