fix(home): Add recent videos, split video_card, add home link
This adds the most recent videos to the home page. As part of this change
we also ensure the top left title links back to the home page. Rendering
for Video cards was moved to an external template and we've switched to
the Jinja environment rendering as recommended.

Some minor formatting tweaks as the Black formatter now works locally.